+++ changelog.md
@@ -30,6 +30,7 @@
 * **[COMMs]** Ability to set a specific callsign to a flight.
 * **[Mission Generator]** Channel terrain fix on exclusion zones, sea zones and inclusion zones
 * **[Options]** Cheat-option for accessing Air Wing Config Dialog after campaign start
+* **[Options]** Option to enable unlimited fuel for AI (player and non-player flights)
 
 ## Fixes
 * **[Mission Generation]** Anti-ship strikes should use "group attack" in their attack-task